+++ lib/modules/theme/theme_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: theme
-// Updated at: 2026-07-04 22:27:42
-// Commit: feat(theme): add async theme stream provider
+// Updated at: 2026-07-05 16:34:14
+// Commit: feat(theme): integrate custom telemetry logger in theme
 
 class themeComponentState {
-  final String commitRef = '2585';
-  final String lastMessage = 'feat(theme): add async theme stream provider';
-  final DateTime timestamp = DateTime.parse('2026-07-04T22:27:42+05:00');
+  final String commitRef = '2605';
+  final String lastMessage = 'feat(theme): integrate custom telemetry logger in theme';
+  final DateTime timestamp = DateTime.parse('2026-07-05T16:34:14+05:00');
 }